major update - ready for DCAT
- Graphics Overhaul
   - Used assets from kenney.nl
   - Overhauled map, enemies, some UI
   - Added water around the island
- Added Enemy Types
   - Zombie, Vampire and Skeleton replace Basic, Fast and Heavy
   - Includes animations
   - Coding in progress
- Added First Turret
   - Removed Test Turret
   - Created model for basic turret
   - Implemented code
- Added Main Menu
- Added Inventory
   - UI Overhaul needed
- Rearranged files

+++ Roguelike Tower Defender/Assets/Scripts/PlayerStats.cs
@@ -2,8 +2,8 @@ using UnityEngine;
 
 public class PlayerStats : MonoBehaviour
 {
-    public static int playerHealth;
-    public int playerStartHealth;
+    public static float playerHealth;
+    public float playerStartHealth;
 
     void Start()
     {
